Implemented the riddle prompt functionality from the quest popup menus. Use:  Menu: ?=Question Reminder to get it. Talk to the Merchant to get this puzzle quest and try it.  "blue" is the correct answer to the riddle.

+++ trunk/src/common/paws/pawslistbox.h
@@ -189,7 +189,7 @@ public:
     void AddRow( pawsListBoxRow* row );
 
     /// Get number of rows in listboz
-    int GetRowCount();
+    size_t GetRowCount();
  
     /// Get a particular row
     pawsListBoxRow* GetRow( size_t x );
